+++ packages/smooth_app/lib/data_models/news_feed/newsfeed_json.dart
@@ -1,6 +1,6 @@
-part of 'tagline_provider.dart';
+part of 'newsfeed_provider.dart';
 
-/// Content from the JSON and converted to what's in "tagmodel.dart"
+/// Content from the JSON and converted to what's in "newsfeed_model.dart"
 
 class _TagLineJSON {
   _TagLineJSON.fromJson(Map<dynamic, dynamic> json)
@@ -15,17 +15,16 @@ class _TagLineJSON {
   final _TagLineJSONNewsList news;
   final _TaglineJSONFeed taglineFeed;
 
-  TagLine toTagLine(String locale) {
-    final Map<String, TagLineNewsItem> tagLineNews = news.map(
-      (String key, _TagLineItemNewsItem value) =>
-          MapEntry<String, TagLineNewsItem>(
+  AppNews toTagLine(String locale) {
+    final Map<String, AppNewsItem> tagLineNews = news.map(
+      (String key, _TagLineItemNewsItem value) => MapEntry<String, AppNewsItem>(
         key,
         value.toTagLineItem(locale),
       ),
     );
 
     final _TagLineJSONFeedLocale localizedFeed = taglineFeed.loadNews(locale);
-    final Iterable<TagLineFeedItem> feed = localizedFeed.news
+    final Iterable<AppNewsFeedItem> feed = localizedFeed.news
         .map((_TagLineJSONFeedLocaleItem item) {
           if (news[item.id] == null) {
             // The asked ID doesn't exist in the news
@@ -33,16 +32,16 @@ class _TagLineJSON {
           }
           return item.overrideNewsItem(news[item.id]!, locale);
         })
-        .where((TagLineFeedItem? item) =>
+        .where((AppNewsFeedItem? item) =>
             item != null &&
             (item.startDate == null ||
                 item.startDate!.isBefore(DateTime.now())) &&
             (item.endDate == null || item.endDate!.isAfter(DateTime.now())))
         .whereNotNull();
 
-    return TagLine(
-      news: TagLineNewsList(tagLineNews),
-      feed: TagLineFeed(
+    return AppNews(
+      news: AppNewsList(tagLineNews),
+      feed: AppNewsFeed(
         feed.toList(growable: false),
       ),
     );
@@ -106,10 +105,10 @@ class _TagLineItemNewsItem {
     return _translations['default']!.merge(translation);
   }
 
-  TagLineNewsItem toTagLineItem(String locale) {
+  AppNewsItem toTagLineItem(String locale) {
     final _TagLineItemNewsTranslation translation = loadTranslation(locale);
     // We can assume the default translation has a non-null title and message
-    return TagLineNewsItem(
+    return AppNewsItem(
       id: id,
       title: translation.title!,
       message: translation.message!,
@@ -224,8 +223,8 @@ class _TagLineNewsImage {
   final double? width;
   final String? alt;
 
-  TagLineImage toTagLineImage() {
-    return TagLineImage(
+  AppNewsImage toTagLineImage() {
+    return AppNewsImage(
       src: url,
       width: width,
       alt: alt,
@@ -303,7 +302,7 @@ class _TagLineNewsStyle {
     );
   }
 
-  TagLineStyle toTagLineStyle() => TagLineStyle.fromHexa(
+  AppNewsStyle toTagLineStyle() => AppNewsStyle.fromHexa(
         titleBackground: titleBackground,
         titleTextColor: titleTextColor,
         titleIndicatorColor: titleIndicatorColor,
@@ -369,7 +368,7 @@ class _TagLineJSONFeedLocaleItem {
   final String id;
   final _TagLineJSONFeedNewsItemOverride? overrideContent;
 
-  TagLineFeedItem overrideNewsItem(
+  AppNewsFeedItem overrideNewsItem(
     _TagLineItemNewsItem newsItem,
     String locale,
   ) {
@@ -384,9 +383,9 @@ class _TagLineJSONFeedLocaleItem {
       );
     }
 
-    final TagLineNewsItem tagLineItem = item.toTagLineItem(locale);
+    final AppNewsItem tagLineItem = item.toTagLineItem(locale);
 
-    return TagLineFeedItem(
+    return AppNewsFeedItem(
       news: tagLineItem,
       startDate: tagLineItem.startDate,
       endDate: tagLineItem.endDate,
