+++ packages/smooth_app/lib/main.dart
@@ -16,9 +16,9 @@ import 'package:provider/single_child_widget.dart';
 import 'package:scanner_shared/scanner_shared.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 import 'package:smooth_app/data_models/continuous_scan_model.dart';
+import 'package:smooth_app/data_models/news_feed/newsfeed_provider.dart';
 import 'package:smooth_app/data_models/preferences/user_preferences.dart';
 import 'package:smooth_app/data_models/product_preferences.dart';
-import 'package:smooth_app/data_models/tagline/tagline_provider.dart';
 import 'package:smooth_app/data_models/user_management_provider.dart';
 import 'package:smooth_app/database/dao_string.dart';
 import 'package:smooth_app/database/local_database.dart';
@@ -224,8 +224,8 @@ class _SmoothAppState extends State<SmoothApp> {
             provide<ContinuousScanModel>(_continuousScanModel),
             provide<PermissionListener>(_permissionListener),
           ],
-          child: ChangeNotifierProvider<TagLineProvider>(
-            create: (BuildContext context) => TagLineProvider(
+          child: ChangeNotifierProvider<AppNewsProvider>(
+            create: (BuildContext context) => AppNewsProvider(
               context.read<UserPreferences>(),
             ),
             lazy: true,
