+++ packages/smooth_app/lib/pages/preferences/user_preferences_language_selector.dart
@@ -3,9 +3,9 @@ import 'package:flutter_gen/gen_l10n/app_localizations.dart';
 import 'package:openfoodfacts/openfoodfacts.dart';
 import 'package:provider/provider.dart';
 import 'package:smooth_app/background/background_task_language_refresh.dart';
+import 'package:smooth_app/data_models/news_feed/newsfeed_provider.dart';
 import 'package:smooth_app/data_models/preferences/user_preferences.dart';
 import 'package:smooth_app/data_models/product_preferences.dart';
-import 'package:smooth_app/data_models/tagline/tagline_provider.dart';
 import 'package:smooth_app/database/local_database.dart';
 import 'package:smooth_app/generic_lib/design_constants.dart';
 import 'package:smooth_app/generic_lib/widgets/language_selector.dart';
@@ -57,9 +57,9 @@ class UserPreferencesLanguageSelector extends StatelessWidget {
               context.read<LocalDatabase>(),
             );
 
-            // Refresh the tagline
+            // Refresh the news feed
             if (context.mounted) {
-              context.read<TagLineProvider>().loadTagLine();
+              context.read<AppNewsProvider>().loadLatestNews();
             }
             // TODO(monsieurtanuki): make it a background task also?
             // no await
